buggy logging, lots stubbed out

+++ Makefile
@@ -8,21 +8,17 @@ CPU = arm7tdmi
 
 CC = arm-none-eabi-gcc
 LD = arm-none-eabi-ld
-AS = $(CC)
+AS = arm-none-eabi-as
 OBJCOPY = arm-none-eabi-objcopy
 
 CFLAGS = -c -g -O0 -Wall -Wextra -Werror -mthumb -mthumb-interwork -mcpu=$(CPU) -mtune=$(CPU) -ffast-math -fomit-frame-pointer -mthumb -mthumb-interwork
 ASFLAGS = -c -mthumb -mthumb-interwork
 LDFLAGS = -g -Map=$(BUILD)/asteroids.map
 
-CC_OBJECTS = \
+AS_OBJECTS = \
 	$(BUILD)/main.o
 
-AS_OBJECTS = \
-	$(BUILD)/entry.o \
-	$(BUILD)/header.o
-
-LDSCRIPT = src/linker.ld
+LDSCRIPT = src/main.ld
 
 
 all: mkbuilddir $(BUILD)/$(TARGET)
@@ -33,13 +29,10 @@ mkbuilddir:
 $(BUILD)/$(TARGET): $(BUILD)/$(TARGET).elf
 	$(OBJCOPY) -O binary $< $@
 
-$(BUILD)/$(TARGET).elf: $(LDSCRIPT) $(CC_OBJECTS) $(AS_OBJECTS)
+$(BUILD)/$(TARGET).elf: $(LDSCRIPT) $(AS_OBJECTS)
 	$(LD) -o $@ $(LDFLAGS) -T$^
 
-$(CC_OBJECTS): $(BUILD)/%.o: src/%.c
-	$(CC) -o $@ $(CFLAGS) $<
-
-$(AS_OBJECTS): $(BUILD)/%.o: src/%.S
+$(AS_OBJECTS): $(BUILD)/%.o: src/%.s
 	$(AS) -o $@ $(ASFLAGS) $<
 
 dockerbuild:
@@ -50,4 +43,4 @@ dockerbuild:
 .PHONY: clean
 clean:
 	rm -rf $(BUILD)
-	docker rmi asteroids-gba/build
+	if docker image inspect asteroids-gba/build >/dev/null 2>&1; then docker rmi asteroids-gba/build; fi
